+++ monarchy-example/src/main/java/com/zhuinden/monarchyexample/Dog.java
@@ -1,5 +1,7 @@
 package com.zhuinden.monarchyexample;
 
+import android.support.v7.util.DiffUtil;
+
 import com.google.auto.value.AutoValue;
 
 /**
@@ -8,6 +10,18 @@ import com.google.auto.value.AutoValue;
 
 @AutoValue
 public abstract class Dog {
+    public static DiffUtil.ItemCallback<Dog> ITEM_CALLBACK = new DiffUtil.ItemCallback<Dog>() {
+        @Override
+        public boolean areItemsTheSame(Dog oldItem, Dog newItem) {
+            return oldItem.name().equals(newItem.name());
+        }
+
+        @Override
+        public boolean areContentsTheSame(Dog oldItem, Dog newItem) {
+            return oldItem.name().equals(newItem.name());
+        }
+    };
+    
     public abstract String name();
 
     public static Dog create(String name) {