+++ monarchy-example/src/main/java/com/zhuinden/monarchyexample/features/home/HomeFragment.java
@@ -14,6 +14,7 @@ import com.zhuinden.monarchyexample.features.copied.CopiedKey;
 import com.zhuinden.monarchyexample.features.managed.ManagedKey;
 import com.zhuinden.monarchyexample.features.mapped.MappedKey;
 import com.zhuinden.monarchyexample.features.mapped_rx.MappedRxKey;
+import com.zhuinden.monarchyexample.features.paged.PagedKey;
 import com.zhuinden.monarchyexample.utils.BaseFragment;
 
 import butterknife.ButterKnife;
@@ -43,6 +44,11 @@ public class HomeFragment
     @OnClick(R.id.button_mapped_rx)
     public void onMappedRx(View view) {
         MainActivity.get(view.getContext()).navigateTo(MappedRxKey.create());
+    }
+
+    @OnClick(R.id.button_paged)
+    public void onPaged(View view) {
+        MainActivity.get(view.getContext()).navigateTo(PagedKey.create());
     }
 
     @Override